requirements.txt
# requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
onnxruntime==1.16.1
numpy==1.24.3
pillow==10.0.1
transformers==4.35.0  # Optional, but useful for some models
scikit-learn==1.3.0
matplotlib==3.7.2
requests==2.31.0